Services folder restructuring, RentalManager ok
