+++ Controllers/OwnerController.cs
@@ -3,12 +3,13 @@ using real_estate_web_api.Models.Entities.People;
 using real_estate_web_api.Models.Results;
 using real_estate_web_api.Models.ViewModels;
 using real_estate_web_api.Services;
+using real_estate_web_api.Services.Owners;
 
 namespace real_estate_web_api.Controllers;
 
 public class OwnerController : StandardController<IOwner, OwnerViewModel, OwnerResult>
 {
-    public OwnerController(IManager<IOwner> manager) : base(manager)
+    public OwnerController(IOwnerManager manager) : base(manager)
     {
     }
