+++ Controllers/RealtorController.cs
@@ -2,12 +2,13 @@ using real_estate_web_api.Models.Entities.People;
 using real_estate_web_api.Models.Results;
 using real_estate_web_api.Models.ViewModels;
 using real_estate_web_api.Services;
+using real_estate_web_api.Services.Realtors;
 
 namespace real_estate_web_api.Controllers;
 
 public class RealtorController : StandardController<IRealtor, RealtorViewModel, RealtorResult>
 {
-    public RealtorController(IManager<IRealtor> manager) : base(manager)
+    public RealtorController(IRealtorManager manager) : base(manager)
     {
     }
 }
