+++ Controllers/RentalController.cs
@@ -2,12 +2,13 @@ using real_estate_web_api.Models.Entities.Rentals;
 using real_estate_web_api.Models.Results;
 using real_estate_web_api.Models.ViewModels;
 using real_estate_web_api.Services;
+using real_estate_web_api.Services.Rentals;
 
 namespace real_estate_web_api.Controllers;
 
 public class RentalController : StandardController<IRental, RentalViewModel, RentalResult>
 {
-    public RentalController(IManager<IRental> manager) : base(manager)
+    public RentalController(IRentalManager manager) : base(manager)
     {
     }
 }
