+++ Services/Owners/IOwnerManager.cs
@@ -1,0 +1,8 @@
+using real_estate_web_api.Models.Entities.People;
+
+namespace real_estate_web_api.Services.Owners;
+
+public interface IOwnerManager : IManager<IOwner>
+{
+
+}
