+++ Services/Owners/OwnerManager.cs
@@ -1,27 +1,27 @@
 using real_estate_web_api.Models.Entities.People;
 
-namespace real_estate_web_api.Services;
+namespace real_estate_web_api.Services.Owners;
 
-public class RealtorManager : IManager<IRealtor>
+public class OwnerManager : IOwnerManager
 {
     private readonly IRepository<Person> _repository;
 
-    public RealtorManager(IRepository<Person> repository)
+    public OwnerManager(IRepository<Person> repository)
     {
         _repository = repository;
     }
 
-    public async Task<ServiceResult<IRealtor>> Create(IRealtor entity)
+    public async Task<ServiceResult<IOwner>> Create(IOwner entity)
     {
         var taxDocumentAvailableResult = await CheckTaxDocument(entity.TaxDocument);
 
         if (!taxDocumentAvailableResult.Success)
         {
             ArgumentNullException.ThrowIfNull(taxDocumentAvailableResult.Error);
-            return new ServiceResult<IRealtor>(taxDocumentAvailableResult.Error);
+            return new ServiceResult<IOwner>(taxDocumentAvailableResult.Error);
         }
 
-        var existingPersonResult = await _repository.Find(x => x.TaxDocument == entity.TaxDocument && !x.IsRealtor);
+        var existingPersonResult = await _repository.Find(x => x.TaxDocument == entity.TaxDocument && !x.IsOwner);
 
         if (existingPersonResult.Success && existingPersonResult.Content != null)
         {
@@ -44,38 +44,38 @@ public class RealtorManager : IManager<IRealtor>
         return result;
     }
 
-    public async Task<ServiceResult<List<IRealtor>>> Retrieve()
+    public async Task<ServiceResult<List<IOwner>>> Retrieve()
     {
-        var result = await _repository.Search(x => x.IsRealtor);
+        var result = await _repository.Search(x => x.IsOwner);
 
         return ToEntityResult(result);
     }
 
-    public async Task<ServiceResult<IRealtor>> Retrieve(string id)
+    public async Task<ServiceResult<IOwner>> Retrieve(string id)
     {
-        var result = await _repository.Find(x => x.IsRealtor && x.Id == id);
+        var result = await _repository.Find(x => x.IsOwner && x.Id == id);
 
         return ToEntityResult(result);
     }
 
-    public async Task<ServiceResult<List<IRealtor>>> Search(Func<IRealtor, bool> filter)
+    public async Task<ServiceResult<List<IOwner>>> Search(Func<IOwner, bool> filter)
     {
-        var result = await _repository.Search(filter, x => x.IsRealtor);
+        var result = await _repository.Search(filter, x => x.IsOwner);
 
         return ToEntityResult(result);
     }
 
-    public async Task<ServiceResult<IRealtor>> Update(IRealtor entity)
+    public async Task<ServiceResult<IOwner>> Update(IOwner entity)
     {
         var taxDocumentAvailable = await CheckTaxDocument(entity.TaxDocument);
 
         if (!taxDocumentAvailable.Success)
         {
             ArgumentNullException.ThrowIfNull(taxDocumentAvailable.Error);
-            return new ServiceResult<IRealtor>(taxDocumentAvailable.Error);
+            return new ServiceResult<IOwner>(taxDocumentAvailable.Error);
         }
 
-        var existingPersonResult = await _repository.Find(x => x.TaxDocument == entity.TaxDocument && !x.IsRealtor);
+        var existingPersonResult = await _repository.Find(x => x.TaxDocument == entity.TaxDocument && !x.IsOwner);
 
         if (existingPersonResult.Success && existingPersonResult.Content != null)
         {
@@ -91,7 +91,7 @@ public class RealtorManager : IManager<IRealtor>
 
     private async Task<ServiceResult> CheckTaxDocument(string taxDocument)
     {
-        var entities = await _repository.Search(x => x.TaxDocument == taxDocument && x.IsRealtor);
+        var entities = await _repository.Search(x => x.TaxDocument == taxDocument && x.IsOwner);
 
         if (!entities.Success || entities.Content == null)
         {
@@ -116,11 +116,11 @@ public class RealtorManager : IManager<IRealtor>
         return new ServiceResult(success: true);
     }
 
-    private Person ToPerson(IRealtor entity)
+    private Person ToPerson(IOwner entity)
     {
         return new Person
         {
-            IsRealtor = true,
+            IsOwner = true,
             Id = entity.Id,
             TaxDocument = entity.TaxDocument,
             Address = entity.Address,
@@ -131,10 +131,9 @@ public class RealtorManager : IManager<IRealtor>
         };
     }
 
-    private void CopyToPerson(IRealtor entity, Person person)
+    private void CopyToPerson(IOwner entity, Person person)
     {
-        person.IsRealtor = true;
-        person.Id = entity.Id;
+        person.IsOwner = true;
         person.TaxDocument = entity.TaxDocument;
         person.Address = entity.Address;
         person.BirthDate = entity.BirthDate;
@@ -143,27 +142,27 @@ public class RealtorManager : IManager<IRealtor>
         person.Mobile = entity.Mobile;
     }
 
-    private ServiceResult<IRealtor> ToEntityResult(ServiceResult<Person> result)
+    private ServiceResult<IOwner> ToEntityResult(ServiceResult<Person> result)
     {
         if (result.Success)
         {
             ArgumentNullException.ThrowIfNull(result.Content);
-            return new ServiceResult<IRealtor>(result.Content);
+            return new ServiceResult<IOwner>(result.Content);
         }
 
         ArgumentNullException.ThrowIfNull(result.Error);
-        return new ServiceResult<IRealtor>(result.Error);
+        return new ServiceResult<IOwner>(result.Error);
     }
 
-    private ServiceResult<List<IRealtor>> ToEntityResult(ServiceResult<List<Person>> result)
+    private ServiceResult<List<IOwner>> ToEntityResult(ServiceResult<List<Person>> result)
     {
         if (result.Success)
         {
             ArgumentNullException.ThrowIfNull(result.Content);
-            return new ServiceResult<List<IRealtor>>(new List<IRealtor>(result.Content));
+            return new ServiceResult<List<IOwner>>(new List<IOwner>(result.Content));
         }
 
         ArgumentNullException.ThrowIfNull(result.Error);
-        return new ServiceResult<List<IRealtor>>(result.Error);
+        return new ServiceResult<List<IOwner>>(result.Error);
     }
 }
