+++ Services/Realtors/IRealtorManager.cs
@@ -1,0 +1,8 @@
+using real_estate_web_api.Models.Entities.People;
+
+namespace real_estate_web_api.Services.Realtors;
+
+public interface IRealtorManager : IManager<IRealtor>
+{
+
+}
