+++ Services/Rentals/IRentalManager.cs
@@ -1,0 +1,8 @@
+using real_estate_web_api.Models.Entities.Rentals;
+
+namespace real_estate_web_api.Services.Rentals;
+
+public interface IRentalManager : IManager<IRental>
+{
+
+}
