+++ Services/Rentals/RentalManager.cs
@@ -1,0 +1,111 @@
+using real_estate_web_api.Models.Entities.RealEstates;
+using real_estate_web_api.Models.Entities.Rentals;
+using real_estate_web_api.Services.Realtors;
+using real_estate_web_api.Services.Tenants;
+
+namespace real_estate_web_api.Services.Rentals;
+
+public class RentalManager : StandardManager<IRental>, IRentalManager
+{
+    private readonly IManager<IRealEstate> _realEstateManager;
+    private readonly IRealtorManager _realtorManager;
+    private readonly ITenantManager _tenantManager;
+
+    public RentalManager(
+        IRepository<IRental> repository,
+        IManager<IRealEstate> realEstateManager,
+        IRealtorManager realtorManager,
+        ITenantManager tentantManager)
+    : base(repository)
+    {
+        _realEstateManager = realEstateManager;
+        _realtorManager = realtorManager;
+        _tenantManager = tentantManager;
+    }
+
+    public override async Task<ServiceResult<IRental>> Create(IRental entity)
+    {
+        var validReferencesResult = await CheckReferences(entity);
+        if (!validReferencesResult.Success)
+            return validReferencesResult;
+
+        var result = await _repository.Create(entity);
+
+        return result;
+    }
+
+    public override async Task<ServiceResult<IRental>> Update(IRental entity)
+    {
+        var result = await _repository.Update(entity);
+
+        return result;
+    }
+
+    private async Task<ServiceResult<IRental>> CheckReferences(IRental entity)
+    {
+        var validRealEstateResult = await CheckRealEstate(entity.RealEstate.Id);
+        if (!validRealEstateResult.Success)
+            return validRealEstateResult;
+
+        var validRealtorResult = await CheckRealtor(entity.Realtor.Id);
+        if (!validRealtorResult.Success)
+            return validRealtorResult;
+
+        var validTenantResult = await CheckTenant(entity.Tenant.Id);
+        if (!validTenantResult.Success)
+            return validTenantResult;
+
+        return new ServiceResult<IRental>(new Rental());
+    }
+
+    private async Task<ServiceResult<IRental>> CheckRealEstate(string id)
+    {
+        var exists = await _realEstateManager.Retrieve(id);
+
+        if (exists.Success && exists.Content != null)
+        {
+            return new ServiceResult<IRental>(new Rental());
+        }
+
+        var error = new ServiceError(
+            "RealEstate not found",
+            $"No RealEstate could be located with id: {id}",
+            404);
+
+        return new ServiceResult<IRental>(error);
+    }
+
+    private async Task<ServiceResult<IRental>> CheckRealtor(string id)
+    {
+        var exists = await _realtorManager.Retrieve(id);
+
+        if (exists.Success && exists.Content != null)
+        {
+            return new ServiceResult<IRental>(new Rental());
+        }
+
+        var error = new ServiceError(
+            "Realtor not found",
+            $"No Realtor could be located with id: {id}",
+            404);
+
+        return new ServiceResult<IRental>(error);
+    }
+
+    private async Task<ServiceResult<IRental>> CheckTenant(string id)
+    {
+        var exists = await _tenantManager.Retrieve(id);
+
+        if (exists.Success && exists.Content != null)
+        {
+            return new ServiceResult<IRental>(new Rental());
+        }
+
+        var error = new ServiceError(
+            "Tenant not found",
+            $"No tenant could be located with id: {id}",
+            404);
+
+        return new ServiceResult<IRental>(error);
+    }
+}
