+++ Services/Tenants/ITenantManager.cs
@@ -1,0 +1,8 @@
+using real_estate_web_api.Models.Entities.People;
+
+namespace real_estate_web_api.Services.Tenants;
+
+public interface ITenantManager : IManager<ITenant>
+{
+
+}
